Usage

Summary

The way that trashing is performed differs wildly between platforms:

On Linux, we trashing works by moving files and creating/reading paired ini-style .trashinfo files (this is the Freedesktop Trashspec).

On MacOS, trashing is done using the Objective-C API, specifically trashItemAtURL:resultingItemURL:error: and manually parsing the .DS_Store file.

On Windows, the Win32 IFileOperation::DeleteItem API is used for trashing, and the remaining functionality is implemented by manually inspecting the recycling bin.

These large differences make a coherent cross-platform API a bit difficult, but by focusing on just supporting basic functionality (trashing and untrashing, listing and emptying trash contents) we have a solid cross-platform trash library.

Example usage

julia> using Trash

julia> write("demofile", "some content")
12

julia> trash("demofile")
TrashFile("/tmp/demofile" @ 2025-04-27)

julia> write("demofile", "more content")
12

julia> trash("demofile")
TrashFile("/tmp/demofile" @ 2025-04-27)

julia> Trash.search("demofile")
2-element Vector{TrashFile}:
 TrashFile("/tmp/demofile" @ 2025-04-27)
 TrashFile("/tmp/demofile" @ 2025-04-27)

julia> untrash("demofile", pick = :oldest)
"demofile"

julia> read("demofile", String)
"some content"

julia> untrash("demofile", force = true)
"demofile"

julia> read("demofile", String)
"more content"

Functions

Trash.trash — Function

trash(path::String; force::Bool=false)

Put the file, link, or empty directory in the system trash. If force=true is passed, a non-existing path is not treated as an error.

See also: Trash.list, untrash.

Trash.untrash — Function

untrash(entry::TrashFile, dest::String = original path;
        force::Bool = false, rm::Bool = false)

Restore a file, link, or directory represented by entry from the system trash.

The entry will be restored to the path dest, which defaults to the original location of the entry.

If force is true, any existing file at the destination will be trashed, and if rm is true, the file will be removed with Base.rm.

See also: trash, Trash.list.

untrash(path::String, dest::String = path; pick::Symbol = :only,
        force::Bool = false, rm::Bool = false)

Restore the original contents of path, optionally specifying a different destination.

The path is the original path of the file or directory to be restored, and has no connection to how the resource is stored in the trash.

Should multiple entries of path exist in the trash, an entry will be chosen based on the pick option. The default is :only, which will throw an ArgumentError if multiple entries are found. The other options are :newest and :oldest, which will select the most recent or oldest entry, respectively.

The force and rm options are passed through to the untrash(::TrashFile) function.

Trash.trashdir — Function

trashdir() -> String

Return the general trash directory for the current user.

trashdir(path::String) -> String

Return the trash directory used for path.

Trash.trashes — Function

trashes() -> Vector{String}

Return a list of all trash directories on the system.

Trash.list — Function

list() -> Vector{TrashFile}

List all entries current in an accessible trash directory.

This searches the system for trash directories on local, mounted, writable filesystems (including removable drives) and combines results into a single list.

Filesystems that are network-based (e.g. NFS, SMB, or SSHFS) are skipped.

See also: trash, untrash, orphans.

list(trashdir::String) -> Vector{TrashFile}

List all entries currently in the trash directory trashdir.

See also: trashdir, search.

Trash.search — Function

search(path::String) -> Vector{TrashFile}

Search for path entries in the trash.

This is a minor convenience function on top of Trash.list, which see.

Trash.orphans — Function

orphans() -> Vector{TrashFile}

List all entries in the trash that are missing data or metadata.

Dangling data produces TrashFile entries with an empty path, while dangling metadata produces full TrashFile entries that cannot be restored.

This searches the system for trash directories on local, mounted, writable filesystems (including removable drives) and combines results into a single list.

Filesystems that are network-based (e.g. NFS, SMB, or SSHFS) are skipped.

See also: Trash.list, Trash.purge.

orphans(trashdir::String) -> Vector{TrashFile}

List all entries in the trash directory trashdir that are missing data or metadata.

Trash.purge — Function

purge(entry::TrashFile)

Permanently delete the entry entry from the trash.

To the extent possible, this will remove the data and metadata associated with the entry, and will not be recoverable.

See also: Trash.list, Trash.orphans, Trash.empty.

Trash.empty — Function

empty()

Empty the user trash.

See also: trash, Trash.list.

empty(trashdir::String)

Empty the trash directory trashdir.

Types

Trash.TrashFile — Type

TrashFile

A representation of a file that has been trashed.

This representation contains:

the path to the trash file,

the path to the original file, and

the time it was trashed

Platform complications

Linux (XDG Trash)

Everything works great 🙂.

MacOS

The "put back" functionality is broken, as the .DS_Store file is not properly modified with the provided system APIs. See https://openradar.appspot.com/23153124 for more information.

We rely on parsing the (undocumented) .DS_Store file to read trash information.

Windows

Everything works, but trashing a file with knowledge of where it ended up and the associated deletion time is rather involved. It's about a thousand times slower to delete a file on Windows than Linux or MacOS, but unless you're wanting to trash thousands and thousands of files, this shouldn't be a concern.

We rely on an undocumented convention for the recycle bin path, as SHGetKnownFolderPath pretends that the recycle bin does not correspond to a physical folder (even though it does). We also rely on the (undocumented) existence of $R... and parsed data from $I... files in the recycle bin to retrieve information on the contents of the recycle bin.